backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt==4.2.1",
    "duckdb==1.1.3",
    "fastapi==0.115.5",
    "passlib[bcrypt]==1.7.4",
    "pydantic==2.10.3",
    "pydantic-settings==2.6.1",
    "python-dotenv==1.0.1",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.18",
    "python-socketio==5.11.4",
    "redis==5.2.0",
    "requests>=2.32.5",
    "uvicorn[standard]==0.32.1",
]

